[USETUP] de-DE Update 2022-03-05 CORE-18075
Avoids some typos like
 gew\201nste -> gew\201nschte
 gew\204hlten -> gew\204hlte

Global replacement of
 fahr->setz
to harmonize the mixed usage of Fortfahren vs Fortsetzen

Also recently someone started using ENTER instead of EINGABETASTE
and although I was pissed off by that in the beginning, after thinking
twice: ENTER is the much better alternative to EINGABETASTE, because:
- it does not need to be translated from english
- it is shorter and therefore helps to avoid some needed abbreviations
- it allows to avoid articles and therefore aligns better with other selections
- it avoids translation errors like EINGABETASTER (and EEINGABETASTE in older branches)
So I use that consistently throughout the translation now.

Also shorten "Der PC wird automatisch neu gestartet" because that line was
so long, that it looked odd.

88 Lines touched: "Der Führer wär stolz"